first commit to v2 of the color library
